Added solution for candidate-Id 30743785
