--- Group 4-Java cluster/30743785/Sports-management/SportsManage.java
@@ -1,32 +1,0 @@
-package cts.sql.sp;
-import java.util.*;
-
-public class SportsManage {
-
-	public static void main(String[] args) throws Exception {
-		Scanner sc = new Scanner(System.in);
-		
-		System.out.println("Welcome to Sports Managemnet");
-		System.out.println("1. Team Table");
-		System.out.println("2. Player Table");
-		System.out.println("3. Match Table");
-		System.out.println("Enter your num to open the table: ");
-		int num = sc.nextInt();
-		Management manage = new Management();
-		switch(num) {
-		case 1:
-			manage.fetchDataTeam();
-			break;
-		case 2 :
-			manage.fetchDataPlayer();
-			break;
-		case 3:
-			manage.fetchDataMatch();
-			break;
-		default:
-			System.out.println("Enter the Correct Details...!");
-			break;
-		
-		}
-	}
-}
