+++ Group 4-Java cluster/30743785/Sports-management/src/com/cts/sql/dao/MatchTable.java
@@ -1,5 +1,8 @@
-package cts.sql.sp;
+package com.cts.sql.dao;
 import java.sql.*;
+
+import com.cts.sql.model.MatchClass;
+
 public class MatchTable {
 
 	public void addMatch(MatchClass matchclass)throws Exception {
