+++ Group 4-Java cluster/30743785/Sports-management/src/com/cts/sql/model/MatchClass.java
@@ -1,4 +1,4 @@
-package cts.sql.sp;
+package com.cts.sql.model;
 
 import java.sql.Date;
