+++ Group 4-Java cluster/30743785/Sports-management/src/com/cts/sql/model/TeamClass.java
@@ -1,4 +1,4 @@
-package cts.sql.sp;
+package com.cts.sql.model;
 public class TeamClass{
     private int teamId;
     private String name;
